extracting stave properties component
