+++ src/modules/documents/index.js
@@ -6,4 +6,10 @@ import type { Question } from 'modules/types'
 export type DocumentsState = {
   questions: List<Question>,
   editing: Question,
+  selectedDescription: string,
+  selectedClef: string,
+  selectedTimeSignature: string,
+  selectedKeySignature: string,
+  selectedMeasures: number,
+  selectedValidators: string,
 }
